refactor: cleanup SDF base class
